requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
google-generativeai==0.8.11
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
pyperclip==1.8.2
python-dotenv==1.0.0
pillow==10.1.0
PyQt5==5.15.10